+++ app/(tabs)/_layout.tsx
@@ -1,9 +1,9 @@
 import FontAwesome from "@expo/vector-icons/FontAwesome";
 import { Link, Tabs } from "expo-router";
-
 import { Pressable, StyleSheet } from "react-native";
-import { Button, ButtonIcon, AddIcon } from "@gluestack-ui/themed";
-import { supabase } from "~/utils/supabase";
+import { Button, ButtonIcon, AddIcon, ButtonText} from "@gluestack-ui/themed";
+import React from "react";
+import GroupsScreen from "./groups";
 
 
 function TabBarIcon(props: {
@@ -13,18 +13,9 @@ function TabBarIcon(props: {
   return <FontAwesome size={28} style={styles.tabBarIcon} {...props} />;
 }
 
-async function createGroup() {
-  console.log("Creating Group")
-  const { data, error } = await supabase
-  .from('groups')
-  .insert([
-    { bio: 'This is a test group creation' },
-  ])
-  .select()
-  console.log(data, error)
-}
 
 export default function TabLayout() {  
+  
   return (
     <Tabs
       screenOptions={{
@@ -60,15 +51,9 @@ export default function TabLayout() {
       /> */}
       <Tabs.Screen
         name="groups"
-        options={{
-          title: "Groups",
+        options= {{
+          title: "Groups", // Custom header title component
           tabBarIcon: ({ color }) => <TabBarIcon name="code" color={color} />,
-          headerRight: () => (
-          <Button borderRadius='$full' size='lg' p='$3.5' bgColor='$warning500' 
-          onPress={() => createGroup()}>
-            <ButtonIcon as={AddIcon}/>
-          </Button>  
-          ),
         }}
       />
       <Tabs.Screen
@@ -87,6 +72,7 @@ export default function TabLayout() {
       />
       <Tabs.Screen
         name="account"
+        
         options={{
           title: "Account",
           tabBarIcon: ({ color }) => <TabBarIcon name="user" color={color} />,
